# Spring-Boot-CRUD-REST-API/SB-REST-API-PostGreSQL-SQLServer/SB-Code/Swagger-Config-Screenshot/crud-REST-api-docs-v1.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ShowCastDB CRUD REST API",
      "description": "A REST API for streamlined management(CRUD) of TV show cast and actor information.",
      "contact": {
        "name": "Ms. Chanandler Bong",
        "email": "ms.chanandler.bong@friends.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "REST CRUD Dev Server"
      }
    ],
    "paths": {
      "/api/showactors/{id}": {
        "get": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "getActorById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ShowActorDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "updateActor",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowActorDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/RESTAPIResponse"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "deleteActor",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/RESTAPIResponse"
                  }
                }
              }
            }
          }
        },
        "patch": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "updateActorShowName",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorShowNameUpdateDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/RESTAPIResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/showactors/": {
        "get": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "getAllActors",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ShowActorDTO"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "show-actor-controller"
          ],
          "operationId": "createActor",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowActorDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/RESTAPIResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "ShowActorDTO": {
          "type": "object",
          "properties": {
            "actorID": {
              "type": "string",
              "format": "uuid"
            },
            "showName": {
              "type": "string"
            },
            "actorShowName": {
              "type": "string"
            },
            "actorRealName": {
              "type": "string"
            }
          }
        },
        "RESTAPIResponse": {
          "type": "object",
          "properties": {
            "httpStatusCode": {
              "type": "integer",
              "format": "int32"
            },
            "responseStatus": {
              "type": "string"
            },
            "responseMessage": {
              "type": "string"
            },
            "timestamp": {
              "type": "string"
            }
          }
        },
        "ActorShowNameUpdateDTO": {
          "type": "object",
          "properties": {
            "actorShowName": {
              "type": "string"
            }
          }
        }
      }
    }
  }